+++ src/c++/perf_analyzer/client_backend/mock_client_backend.h
@@ -27,6 +27,7 @@
 #pragma once
 
 #include <chrono>
+#include <mutex>
 #include "client_backend.h"
 
 namespace triton { namespace perfanalyzer { namespace clientbackend {
